refactor(vats): factor out promise-space.js
