+++ packages/vats/test/test-name-hub-published.js
@@ -4,7 +4,8 @@ import { test } from '@agoric/swingset-vat/tools/prepare-test-env-ava.js';
 import { makeMockChainStorageRoot } from '@agoric/inter-protocol/test/supports.js';
 import { makeHandle } from '@agoric/zoe/src/makeHandle.js';
 import { eventLoopIteration } from '@agoric/zoe/tools/eventLoopIteration.js';
-import { makeAgoricNamesAccess, makePromiseSpace } from '../src/core/utils.js';
+import { makeAgoricNamesAccess } from '../src/core/utils.js';
+import { makePromiseSpace } from '../src/core/promise-space.js';
 import {
   publishAgoricNames,
   setupClientManager,
